+++ CHANGELOG.md
@@ -9,12 +9,14 @@ All notable changes to this project will be documented in this file.
 - a changelog
 - Display a stacktrace entry that leads to the test method for failed tests.
 - Report errors in contexts as failing tests.
+- Tests are reported in the same order as they appear in the file.
+- Contexts are reported in the same order as they appear in the file. who would have thought that they were reported in
+  random order before.
 
 ### Changed
 
 - Duplicate contexts fail the suite.
 - Contexts with the same name as a test in the same context fail the suite.
-- Tests are reported in the same order as they appear in the file.
 
 ### Fixed